+++ src/platform/macos/events_loop.rs
@@ -442,8 +442,8 @@ impl EventsLoop {
             appkit::NSOtherMouseDown => { Some(into_event(WindowEvent::MouseInput { device_id: DEVICE_ID, state: ElementState::Pressed, button: MouseButton::Middle })) },
             appkit::NSOtherMouseUp => { Some(into_event(WindowEvent::MouseInput { device_id: DEVICE_ID, state: ElementState::Released, button: MouseButton::Middle })) },
 
-            appkit::NSMouseEntered => { Some(into_event(WindowEvent::MouseEntered { device_id: DEVICE_ID })) },
-            appkit::NSMouseExited => { Some(into_event(WindowEvent::MouseLeft { device_id: DEVICE_ID })) },
+            appkit::NSMouseEntered => { Some(into_event(WindowEvent::CursorEntered { device_id: DEVICE_ID })) },
+            appkit::NSMouseExited => { Some(into_event(WindowEvent::CursorLeft { device_id: DEVICE_ID })) },
 
             appkit::NSMouseMoved |
             appkit::NSLeftMouseDragged |
@@ -474,7 +474,7 @@ impl EventsLoop {
                 {
                     let x = (scale_factor * view_point.x as f32) as f64;
                     let y = (scale_factor * (view_rect.size.height - view_point.y) as f32) as f64;
-                    let window_event = WindowEvent::MouseMoved { device_id: DEVICE_ID, position: (x, y) };
+                    let window_event = WindowEvent::CursorMoved { device_id: DEVICE_ID, position: (x, y) };
                     let event = Event::WindowEvent { window_id: ::WindowId(window.id()), event: window_event };
                     events.push_back(event);
                 }
@@ -489,6 +489,12 @@ impl EventsLoop {
                 let delta_y = (scale_factor * ns_event.deltaY() as f32) as f64;
                 if delta_y != 0.0 {
                     let motion_event = DeviceEvent::Motion { axis: 1, value: delta_y };
+                    let event = Event::DeviceEvent{ device_id: DEVICE_ID, event: motion_event };
+                    events.push_back(event);
+                }
+
+                if delta_x != 0.0 || delta_y != 0.0 {
+                    let motion_event = DeviceEvent::MouseMotion { delta: (delta_x, delta_y) };
                     let event = Event::DeviceEvent{ device_id: DEVICE_ID, event: motion_event };
                     events.push_back(event);
                 }
@@ -519,7 +525,19 @@ impl EventsLoop {
                     NSEventPhase::NSEventPhaseEnded => TouchPhase::Ended,
                     _ => TouchPhase::Moved,
                 };
-                let window_event = WindowEvent::MouseWheel { device_id: DEVICE_ID, delta: delta, phase: phase };
+                self.shared.pending_events.lock().unwrap().push_back(Event::DeviceEvent {
+                    device_id: DEVICE_ID,
+                    event: DeviceEvent::MouseWheel {
+                        delta: if ns_event.hasPreciseScrollingDeltas() == cocoa::base::YES {
+                            PixelDelta(ns_event.scrollingDeltaX() as f32,
+                                       ns_event.scrollingDeltaY() as f32)
+                        } else {
+                            LineDelta(ns_event.scrollingDeltaX() as f32,
+                                      ns_event.scrollingDeltaY() as f32)
+                        },
+                    }
+                });
+                let window_event = WindowEvent::MouseWheel { device_id: DEVICE_ID, delta, phase };
                 Some(into_event(window_event))
             },